feat: allow custom rtp files
Currently, the files must be placed **inside** a directory named in
accordance to the output directory in which the user wishes their file /
folder to be available (ex: `lua`, `parser`, `after`, etc.).

This could possibly be done more cleanly by making the `rtp` option a
map (attribute set) of paths and then manually stitching the resulting
derivations back together using `symlinkJoin`, but for now this is
simpler while still getting the job done.