feat: add convenience function for simple modules
This function basically just abstracts away the all of the
required boilerplate for a module. Especially useful for simple plugins
which only need to create a `enable` and `opts` option, and then
conditionally add themselves to the plugin spec in `config`. There is
also quite a lot of leeway via the `extraOptions`, etc. args, enabling
more "complex" plugins / modules to still be constructed via this method.

Of course, this function just returns a normal module attribute set, so
the conventional approach is equally valid and supported.